requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysqlclient==2.2.1
numpy==1.26.0
pandas==2.1.1
Pillow==10.1.0
PyJWT==2.8.0
requests==2.31.0
resend==0.6.0
rest-framework-simplejwt==0.0.2
scikit-learn==1.3.1
scipy==1.11.3
python==3.10.9
channels==4.0.0
gunicorn==21.2.0npm star

moviepy~=1.0.3
opencv-python~=4.8.1.78
django-simple-history~=3.4.0
waitress~=2.1.2
django-cors-headers==3.13.0